+++ Example/Stopwatch/AppDelegate.h
@@ -1,7 +1,8 @@
 #import <UIKit/UIKit.h>
 #import "ABKPushURIDelegate.h"
+#import <UserNotifications/UserNotifications.h>
 
-@interface AppDelegate : UIResponder <UIApplicationDelegate, ABKPushURIDelegate>
+@interface AppDelegate : UIResponder <UIApplicationDelegate, ABKPushURIDelegate, UNUserNotificationCenterDelegate>
 
 @property (strong, nonatomic) UIWindow *window;