+++ HelloSwift/HelloSwift/AppDelegate.swift
@@ -1,38 +1,37 @@
 import UIKit
 import Appboy_iOS_SDK
+import UserNotifications
 
 @UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
+class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterDelegate {
 
   var window: UIWindow?
 
-  func application(application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject: AnyObject]?) -> Bool {
+  func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey : Any]? = nil) -> Bool {
     // Override point for customization after application launch.
-    Appboy.startWithApiKey("6deb7788-edee-4a02-b75a-a254cdb9f58b", inApplication:application, withLaunchOptions:launchOptions)
+    Appboy.start(withApiKey: "6deb7788-edee-4a02-b75a-a254cdb9f58b", in:application, withLaunchOptions:launchOptions)
     
-    // Register push
-    let setting : UIUserNotificationSettings = UIUserNotificationSettings(forTypes: [.Badge, .Sound, .Alert], categories: nil)
-    UIApplication.sharedApplication().registerUserNotificationSettings(setting)
-    UIApplication.sharedApplication().registerForRemoteNotifications()
+    let center = UNUserNotificationCenter.current()
+    center.requestAuthorization(options: [.alert, .badge, .sound]) { (granted, error) in
+      print("Permission granted.")
+    }
+    center.delegate = self
+    UIApplication.shared.registerForRemoteNotifications()
     
     return true
   }
 
-  func applicationDidBecomeActive(application: UIApplication) {
-    UIApplication.sharedApplication().applicationIconBadgeNumber = 0
+  func applicationDidBecomeActive(_ application: UIApplication) {
+    UIApplication.shared.applicationIconBadgeNumber = 0
   }
   
-  func application(application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: NSData) {
-    let deviceTokenString = String(format: "%@", deviceToken)
+  func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
+    let deviceTokenString = String(format: "%@", deviceToken.description)
     Appboy.sharedInstance()!.registerPushToken(deviceTokenString)
   }
   
-  func application(application: UIApplication, didReceiveRemoteNotification userInfo: [NSObject : AnyObject]) {
-    Appboy.sharedInstance()!.registerApplication(application, didReceiveRemoteNotification: userInfo)
-  }
-  
-  func application(application: UIApplication, didReceiveRemoteNotification userInfo: [NSObject : AnyObject], fetchCompletionHandler completionHandler: (UIBackgroundFetchResult) -> Void) {
-    Appboy.sharedInstance()!.registerApplication(application, didReceiveRemoteNotification: userInfo, fetchCompletionHandler: completionHandler)
+  func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void) {
+    Appboy.sharedInstance()!.userNotificationCenter(center, didReceive: response, withCompletionHandler: completionHandler)
   }
 }